refactor EditorUI, refactor diff, wip on FileTreeView

+++ demo-edit-es-module/src/main/java/org/sudu/experiments/JsCodeEditor0.java
@@ -72,7 +72,7 @@ public class JsCodeEditor0 implements JsCodeEditor {
   }
 
   @Override
-  public void setFontSize(int fontSize) {
+  public void setFontSize(float fontSize) {
     editor.changeFont(editor.getFontFamily(), fontSize);
   }
